[oops] extract check-student worker
